requirements.txt
flask==2.3.3
twilio==8.10.0
python-dotenv==1.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.2
pandas==2.1.0
APScheduler==3.10.4
python-dateutil==2.8.2 